Move email utilities to Nest.js service
related to https://github.com/givepraise/praise/pull/1099#issuecomment-1614556721
