+++ packages/api/src/email/templates/EmailTemplateInterface.ts
