--- packages/api/src/shared/email/templates/communityCreatedTemplate.ts
@@ -1,8 +1,0 @@
-import { EmailTemplate } from './EmailTemplateInterface';
-
-export const communityCreatedTemplate: EmailTemplate = {
-  subject: 'Congratulation for creatinf community',
-  html: (payload: any) => `
-    <h1>Congratulation for creating, ${payload.communityName} community!</h1>
-  `,
-};
